+++ src/racer/codeiter.rs
@@ -1,15 +1,18 @@
-use core::{Src, CodeChunkIter};
-
+use codecleaner;
 use std::iter::{Fuse, Iterator};
 
-pub struct StmtIndicesIter<'a> {
-    src: Src<'a>,
-    it: CodeChunkIter<'a>,
+pub struct StmtIndicesIter<'a,I> 
+    where I: Iterator<Item=(usize,usize)>
+{
+    src: &'a str,
+    it: I,
     pos: usize,
     end: usize
 }
 
-impl<'a> Iterator for StmtIndicesIter<'a> {
+impl<'a,I> Iterator for StmtIndicesIter<'a,I> 
+    where I: Iterator<Item=(usize,usize)>
+{
     type Item = (usize, usize);
 
     #[inline]
@@ -114,9 +117,19 @@ fn is_a_use_stmt(src_bytes: &[u8], start: usize, pos: usize) -> bool {
      whitespace.contains(&src_bytes[start+7]))
 }
 
-pub fn iter_stmts(src: Src) -> Fuse<StmtIndicesIter> {
-    StmtIndicesIter{ src: src, it: src.chunk_indices(), pos: 0, end: 0 }.fuse()
+pub fn iter_stmts(src: &str) -> Fuse<StmtIndicesIter<codecleaner::CodeIndicesIter>> {
+    let it = codecleaner::code_chunks(src);
+    StmtIndicesIter{ src: src, it: it, pos: 0, end: 0 }.fuse()
 }
+
+impl<'a,I> StmtIndicesIter<'a,I> 
+        where I: Iterator<Item=(usize,usize)> {
+
+    pub fn from_parts(src: &str, it: I) -> Fuse<StmtIndicesIter<I>> {
+        StmtIndicesIter{ src: src, it: it, pos: 0, end: 0 }.fuse()
+    }
+}
+
 
 #[cfg(test)]
 mod test {
@@ -125,7 +138,7 @@ mod test {
 
     #[test]
     fn iterates_single_use_stmts() {
-        let src = test_source!(rejustify "
+        let src = rejustify("
             use std::Foo; // a comment
             use std::Bar;
         ");
@@ -137,7 +150,7 @@ mod test {
 
     #[test]
     fn iterates_use_stmt_over_two_lines() {
-        let src = test_source!(rejustify "
+        let src = rejustify("
         use std::{Foo,
                   Bar}; // a comment
         ");
@@ -148,7 +161,7 @@ mod test {
 
     #[test]
     fn iterates_use_stmt_without_the_prefix() {
-        let src = test_source!(rejustify "
+        let src = rejustify("
         pub use {Foo,
                  Bar}; // this is also legit apparently
         ");
@@ -160,7 +173,7 @@ mod test {
 
     #[test]
     fn iterates_while_stmt() {
-        let src = test_source!(rejustify "
+        let src = rejustify("
             while self.pos < 3 { }
         ");
         let mut it = iter_stmts(src.as_ref());
@@ -169,7 +182,7 @@ mod test {
 
     #[test]
     fn iterates_lambda_arg() {
-        let src = test_source!(rejustify "
+        let src = rejustify("
             myfn(|n|{});
         ");
         let mut it = iter_stmts(src.as_ref());
@@ -178,13 +191,13 @@ mod test {
 
     #[test]
     fn iterates_macro() {
-        let src = test_source!("
+        let src = "
         mod foo;
         macro_rules! otry(
             ($e:expr) => (match $e { Some(e) => e, None => return })
         )
         mod bar;
-        ");
+        ";
         let mut it = iter_stmts(src.as_ref());
         assert_eq!("mod foo;", slice(&src, it.next().unwrap()));
         assert_eq!("macro_rules! otry(
@@ -195,11 +208,11 @@ mod test {
 
     #[test]
     fn iterates_macro_invocation() {
-        let src = test_source!("
+        let src = "
             mod foo;
             local_data_key!(local_stdout: Box<Writer + Send>)  // no ';'
             mod bar;
-        ");
+        ";
         let mut it = iter_stmts(src.as_ref());
         assert_eq!("mod foo;", slice(&src, it.next().unwrap()));
         assert_eq!("local_data_key!(local_stdout: Box<Writer + Send>)", slice(&src, it.next().unwrap()));
@@ -208,9 +221,9 @@ mod test {
 
     #[test]
     fn iterates_if_else_stmt() {
-        let src = test_source!("
+        let src = "
             if self.pos < 3 { } else { }
-        ");
+        ";
         let mut it = iter_stmts(src.as_ref());
         assert_eq!("if self.pos < 3 { }", slice(&src, it.next().unwrap()));
         assert_eq!("else { }", slice(&src, it.next().unwrap()));
@@ -218,15 +231,15 @@ mod test {
 
     #[test]
     fn iterates_inner_scope() {
-        let src = test_source!(&"
-        while self.pos < 3 {
+        let src = &"
+        while(self.pos < 3 {
             let a = 35;
             return a + 35;  // should iterate this
         }
         {
             b = foo;       // but not this
         }
-        "[29..]);
+        "[29..];
 
         let mut it = iter_stmts(src.as_ref());
 
@@ -237,7 +250,7 @@ mod test {
 
     #[test]
     fn iterates_module_attribute() {
-        let src = test_source!(rejustify "
+        let src = rejustify("
             #![license = \"BSD\"]
             #[test]
         ");
@@ -248,12 +261,12 @@ mod test {
 
     #[test]
     fn iterates_half_open_subscope_if_is_the_last_thing() {
-        let src = test_source!("
+        let src = "
             let something = 35;
             while self.pos < 3 {
             let a = 35;
             return a + 35;  // should iterate this
-        ");
+        ";
 
         let mut it = iter_stmts(src.as_ref());
         assert_eq!("let something = 35;", slice(&src, it.next().unwrap()));