+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][unreleased]
 ### Changed
+ - Add some caching of file source and code indices
+
  - Fix issue resolving some 	std::* modules in latest rust source: (rust std lib implicitly imports core with #![no_std])
 
  - Update emacs racer.el to write temporary files to tmpdir rather than the same directory as the source file